third_party/WebKit/LayoutTests/inspector/tracing/timeline-aggregated-details.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../tracing-test.js"></script>
<script>
function initialize_EventsView()
{
    Runtime.experiments.enableForTest("timelineEventsTreeView");
}

function test()
{
    var sessionId = "6.23";
    var rawTraceEvents = [
    {
        "args": {
          "name": "Renderer"
        },
        "cat": "__metadata",
        "name": "process_name",
        "ph": "M",
        "pid": 17851,
        "tid": 23,
        "ts": 0
    },
    {
        "args": {
            "name": "CrRendererMain"
        },
        "cat": "__metadata",
        "name": "thread_name",
        "ph": "M",
        "pid": 17851,
        "tid": 23,
        "ts": 0
    },
    {
        "args": {
            "sessionId": sessionId
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "TracingStartedInPage",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 100000,
        "tts": 606543
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "Program",
        "ph": "B",
        "pid": 17851,
        "tid": 23,
        "ts": 200000,
        "tts": 5612442
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "c", "callUID": "c", "scriptId": 1 },
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 208000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "FunctionCall",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 210000,
        "dur": 30000,
        "tts": 5612442
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "c", "callUID": "c", "scriptId": 1 },
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 211000,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": []
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 212000,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "c", "callUID": "c", "scriptId": 1 },
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 219875,
        "tts": 1758056
    },
    {
        "args": {
            "frame": "0x2f7b63884000",
            "data": {
                "stackTrace": [
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "InvalidateLayout",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 220000,
        "dur": 7000,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "c", "callUID": "c", "scriptId": 1 },
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 220125,
        "tts": 1758056
    },
    {
        "args": {
            "frame": "0x2f7b63884000",
            "data": {
                "stackTrace": [
                    { "functionName": "b", "callUID": "b", "scriptId": 1 },
                    { "functionName": "a", "callUID": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "InvalidateLayout",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 221000,
        "dur": 3000,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "g", "scriptId": 1 },
                    { "functionName": "f", "scriptId": 1 },
                    { "functionName": "b", "scriptId": 1 },
                    { "functionName": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 222000,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "g", "scriptId": 1 },
                    { "functionName": "e", "scriptId": 1 },
                    { "functionName": "b", "scriptId": 1 },
                    { "functionName": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227125,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "g", "scriptId": 1 },
                    { "functionName": "e", "scriptId": 1 },
                    { "functionName": "b", "scriptId": 1 },
                    { "functionName": "a", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227250,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "a", "callUID": "a", "scriptId": 1 },
                    { "functionName": "l", "callUID": "l", "scriptId": 1 },
                    { "functionName": "f", "callUID": "f", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230000,
        "tts": 1758056
    },
    {
        "args": {
            "beginData": {
                "stackTrace": [
                    { "functionName": "a", "callUID": "a", "scriptId": 1 },
                    { "functionName": "l", "callUID": "l", "scriptId": 1 },
                    { "functionName": "f", "callUID": "f", "scriptId": 1 },
                    { "functionName": "sin", "callUID": "sin", "scriptId": 2, "url": "native math.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "Layout",
        "ph": "X",
        "dur": 100,
        "pid": 17851,
        "tid": 23,
        "ts": 230010,
        "tts": 1758056
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "a", "callUID": "a", "scriptId": 1 },
                    { "functionName": "l", "callUID": "l", "scriptId": 1 },
                    { "functionName": "f", "callUID": "f", "scriptId": 1 },
                    { "functionName": "sin", "callUID": "sin", "scriptId": 2, "url": "native math.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "TimerInstall",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230111
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "a", "callUID": "a", "scriptId": 1 },
                    { "functionName": "l", "callUID": "l", "scriptId": 1 },
                    { "functionName": "f", "callUID": "f", "scriptId": 1 },
                    { "functionName": "sin", "callUID": "sin", "scriptId": 2, "url": "native math.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230125
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "FunctionCall",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 250000,
        "dur": 10000
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "FunctionCall",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 251000,
        "dur": 1000
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "w", "callUID": "w", "scriptId": 1 },
                    { "functionName": "z", "callUID": "z", "scriptId": 1 },
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 251000
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "w", "callUID": "w", "scriptId": 1 },
                    { "functionName": "z", "callUID": "z", "scriptId": 1 },
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 251100
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "w", "scriptId": 1 },
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 251200
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "w", "scriptId": 1 },
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 251300
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "y", "callUID": "y", "scriptId": 1 },
                    { "functionName": "x", "callUID": "x", "scriptId": 1 }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 251400
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_b", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_b", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253000
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_b", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253008
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_b", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253012
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253014
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253015
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_b", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" },
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253100
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                    { "functionName": "recursive_a", "scriptId": 1, "url": "http://xyz.abc.com/rec.js" }
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253200
    },
    {
        "args": {
            "data": {
                "stackTrace": [
                ]
            }
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 253300
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "Program",
        "ph": "E",
        "pid": 17851,
        "tid": 23,
        "ts": 500000,
        "tts": 5612506
    }
    ];

    var timeline = WebInspector.panels.timeline;
    timeline._model.addEventListener(WebInspector.TimelineModel.Events.RecordingStopped, timelineRecorded, this);
    timeline._model.setEventsForTest(rawTraceEvents);

    function timelineRecorded()
    {
        timeline.requestWindowTimes(0, Infinity);
        for (var grouping in WebInspector.AggregatedTimelineTreeView.GroupBy) {
            var groupingValue = WebInspector.AggregatedTimelineTreeView.GroupBy[grouping];
            testEventTree(WebInspector.TimelinePanel.DetailsTab.CallTree, groupingValue);
            testEventTree(WebInspector.TimelinePanel.DetailsTab.BottomUp, groupingValue);
        }

        testEventTree(WebInspector.TimelinePanel.DetailsTab.Events);
        InspectorTest.completeTest();
    }

    function testEventTree(type, grouping)
    {
        InspectorTest.addResult("");
        timeline._detailsView.selectTab(type, true);
        var callTree = timeline._detailsView._rangeDetailViews.get(type);
        if (grouping) {
            InspectorTest.addResult(type + "  Group by: " + grouping);
            callTree._groupByCombobox.select(callTree._groupByCombobox.options().find((x) => x.value === grouping));
            callTree._onGroupByChanged();
        } else {
            InspectorTest.addResult(type);
        }
        var rootNode = callTree._dataGrid.rootNode();
        for (var node of rootNode.children)
            printEventTree(1, node._profileNode);
    }

    function printEventTree(padding, node)
    {
        var name = node.name || (node.event.name === WebInspector.TimelineModel.RecordType.JSFrame
            ? WebInspector.beautifyFunctionName(node.event.args["data"]["functionName"])
            : WebInspector.TimelineUIUtils.eventTitle(node.event));
        InspectorTest.addResult("  ".repeat(padding) + name + ": " + [node.selfTime, node.totalTime].map(function (t) { return t.toFixed(3); }).join("  "));
        (node.children || new Map()).forEach(printEventTree.bind(null, padding + 1));
    }
}

</script>
</head>
<body onload="runTest()">
</body>
</html>
