third_party/WebKit/LayoutTests/inspector/tracing/timeline-aggregated-details-expected.txt


CallTree  Group by: None
  a: 0.000  8.380
    b: 7.125  8.380
      c: 0.351  0.351
      e: 0.000  0.905
        g: 0.905  0.905
  x: 0.000  1.513
    y: 1.113  1.513
      z: 0.000  0.200
        w: 0.200  0.200
      w: 0.200  0.200
  f: 0.000  0.833
    l: 0.000  0.833
      a: 0.733  0.833
        Layout: 0.100  0.100
  recursive_a: 0.101  0.215
    recursive_b: 0.102  0.114
      recursive_a: 0.004  0.012
        recursive_b: 0.008  0.008

BottomUp  Group by: None
  b: 7.125  8.380
    a: 7.125  8.380
  y: 1.113  1.513
    x: 1.113  1.513
  g: 0.905  0.905
    e: 0.905  0.905
      b: 0.905  0.905
        a: 0.905  0.905
  a: 0.733  9.213
    l: 0.733  0.833
      f: 0.733  0.833
  w: 0.400  0.400
    z: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.200  0.200
      x: 0.200  0.200
  c: 0.351  0.351
    b: 0.351  0.351
      a: 0.351  0.351
  recursive_b: 0.110  0.114
    recursive_a: 0.110  0.114
      recursive_b: 0.008  0.008
        recursive_a: 0.008  0.008
  recursive_a: 0.105  0.215
    recursive_b: 0.004  0.012
      recursive_a: 0.004  0.012
  Layout: 0.100  0.100
    a: 0.100  0.100
      l: 0.100  0.100
        f: 0.100  0.100

CallTree  Group by: Category
  Scripting: 0.101  10.941
    a: 0.000  8.380
      b: 7.125  8.380
        c: 0.351  0.351
        e: 0.000  0.905
          g: 0.905  0.905
    f: 0.000  0.833
      l: 0.000  0.833
        a: 0.733  0.833
          Layout: 0.100  0.100
    x: 0.000  1.513
      y: 1.113  1.513
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Category
  Scripting: 10.841  10.841
    a: 0.733  9.213
      l: 0.733  0.833
        f: 0.733  0.833
    b: 7.125  8.380
      a: 7.125  8.380
    c: 0.351  0.351
      b: 0.351  0.351
        a: 0.351  0.351
    e: 0.000  0.905
      b: 0.000  0.905
        a: 0.000  0.905
    g: 0.905  0.905
      e: 0.905  0.905
        b: 0.905  0.905
          a: 0.905  0.905
    f: 0.000  0.833
    l: 0.000  0.833
      f: 0.000  0.833
    x: 0.000  1.513
    y: 1.113  1.513
      x: 1.113  1.513
    z: 0.000  0.200
      y: 0.000  0.200
        x: 0.000  0.200
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
  Rendering: 0.100  0.100
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100

CallTree  Group by: Domain
  unattributed: 0.000  10.726
    a: 0.000  8.380
      b: 7.125  8.380
        c: 0.351  0.351
        e: 0.000  0.905
          g: 0.905  0.905
    f: 0.000  0.833
      l: 0.000  0.833
        a: 0.733  0.833
          Layout: 0.100  0.100
    x: 0.000  1.513
      y: 1.113  1.513
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  abc.com: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Domain
  unattributed: 10.726  10.726
    a: 0.733  9.213
      l: 0.733  0.833
        f: 0.733  0.833
    b: 7.125  8.380
      a: 7.125  8.380
    c: 0.351  0.351
      b: 0.351  0.351
        a: 0.351  0.351
    e: 0.000  0.905
      b: 0.000  0.905
        a: 0.000  0.905
    g: 0.905  0.905
      e: 0.905  0.905
        b: 0.905  0.905
          a: 0.905  0.905
    f: 0.000  0.833
    l: 0.000  0.833
      f: 0.000  0.833
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    x: 0.000  1.513
    y: 1.113  1.513
      x: 1.113  1.513
    z: 0.000  0.200
      y: 0.000  0.200
        x: 0.000  0.200
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
  abc.com: 0.215  0.215
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008

CallTree  Group by: Subdomain
  unattributed: 0.000  10.726
    a: 0.000  8.380
      b: 7.125  8.380
        c: 0.351  0.351
        e: 0.000  0.905
          g: 0.905  0.905
    f: 0.000  0.833
      l: 0.000  0.833
        a: 0.733  0.833
          Layout: 0.100  0.100
    x: 0.000  1.513
      y: 1.113  1.513
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  xyz.abc.com: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Subdomain
  unattributed: 10.726  10.726
    a: 0.733  9.213
      l: 0.733  0.833
        f: 0.733  0.833
    b: 7.125  8.380
      a: 7.125  8.380
    c: 0.351  0.351
      b: 0.351  0.351
        a: 0.351  0.351
    e: 0.000  0.905
      b: 0.000  0.905
        a: 0.000  0.905
    g: 0.905  0.905
      e: 0.905  0.905
        b: 0.905  0.905
          a: 0.905  0.905
    f: 0.000  0.833
    l: 0.000  0.833
      f: 0.000  0.833
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    x: 0.000  1.513
    y: 1.113  1.513
      x: 1.113  1.513
    z: 0.000  0.200
      y: 0.000  0.200
        x: 0.000  0.200
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
  xyz.abc.com: 0.215  0.215
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008

CallTree  Group by: URL
  unattributed: 0.000  10.726
    a: 0.000  8.380
      b: 7.125  8.380
        c: 0.351  0.351
        e: 0.000  0.905
          g: 0.905  0.905
    f: 0.000  0.833
      l: 0.000  0.833
        a: 0.733  0.833
          Layout: 0.100  0.100
    x: 0.000  1.513
      y: 1.113  1.513
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  http://xyz.abc.com/rec.js: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: URL
  unattributed: 10.726  10.726
    a: 0.733  9.213
      l: 0.733  0.833
        f: 0.733  0.833
    b: 7.125  8.380
      a: 7.125  8.380
    c: 0.351  0.351
      b: 0.351  0.351
        a: 0.351  0.351
    e: 0.000  0.905
      b: 0.000  0.905
        a: 0.000  0.905
    g: 0.905  0.905
      e: 0.905  0.905
        b: 0.905  0.905
          a: 0.905  0.905
    f: 0.000  0.833
    l: 0.000  0.833
      f: 0.000  0.833
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    x: 0.000  1.513
    y: 1.113  1.513
      x: 1.113  1.513
    z: 0.000  0.200
      y: 0.000  0.200
        x: 0.000  0.200
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
  http://xyz.abc.com/rec.js: 0.215  0.215
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008

Events
  a: 0.000  0.226
    b: 0.000  0.226
      c: 0.226  0.226
  a: 0.000  8.155
    b: 7.125  8.155
      c: 0.125  0.125
      e: 0.000  0.905
        g: 0.905  0.905
  f: 0.000  0.833
    l: 0.000  0.833
      a: 0.733  0.833
        Layout: 0.100  0.100
        Install Timer: 0.000  0.000
  x: 0.000  1.513
    y: 1.113  1.513
      z: 0.000  0.200
        w: 0.200  0.200
      w: 0.200  0.200
  recursive_a: 0.001  0.015
    recursive_b: 0.002  0.014
      recursive_a: 0.004  0.012
        recursive_b: 0.008  0.008
  recursive_a: 0.100  0.200
    recursive_b: 0.100  0.100
